feat(FavoritesListPage): add FavoritesListPage and extract FileIndexBody
