+++ lib/file/presentation/view/view_export.dart
@@ -1,0 +1,4 @@
+export 'favorites_list_page.dart';
+export 'file_index_page.dart';
+export 'name_file_page.dart';
+export 'photo_detail_page.dart';
